+++ R/Reduce.R
@@ -37,6 +37,7 @@ NULL
 
 reduceexpand <- function(f, args,
                    env = parent.frame()) {
+  force(env)
   fnam <- wrapr_deparse(f)
   if(is.name(f) || is.character(f)) {
     # name of function case as in c(1, 2, 3) %.|% sum